+++ cmd/function_build.go
@@ -81,7 +81,7 @@ func functionBuild(faasCliPath, flagYaml string) error {
 	}
 
 	if res.ExitCode != 0 {
-		return fmt.Errorf("exit code %d", res.ExitCode)
+		return fmt.Errorf(res.Stderr)
 	}
 
 	return nil
@@ -102,7 +102,7 @@ func functionPush(faasCliPath, flagYaml string) error {
 	}
 
 	if res.ExitCode != 0 {
-		return fmt.Errorf("exit code %d", res.ExitCode)
+		return fmt.Errorf(res.Stderr)
 	}
 
 	return nil