+++ cmd/function_deploy.go
@@ -79,7 +79,7 @@ func functionDeploy(kubectlPath string) error {
 	}
 
 	if res.ExitCode != 0 {
-		return fmt.Errorf("exit code %d", res.ExitCode)
+		return fmt.Errorf(res.Stderr)
 	}
 
 	return nil