+++ CHANGELOG.md
@@ -1,14 +1,20 @@
 # Release Notes
 
+## Version v0.0.1 - 2020-05-29
+Improvements:
+    
+- Add check run root in `install.sh`
+- Improvements in all errors messages
+
 ## Version v0.0.1-beta - 2020-05-26
 Features:
     
 - Add new command `infra secrets` to get user and password of applications
-- `function new` Add file .env
+- `function new` Add a file .env
 
 Improvements:
     
-- `function new` Add folder deploy in .gitignore
+- `function new` Add a folder deploy in .gitignore
 
 ## Version v0.0.1-alpha - 2020-05-25
 Initial Safira release